backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# AI/ML Services
google-generativeai==0.3.2
google-genai==1.54.0
azure-cognitiveservices-speech==1.35.0
openai==1.10.0
gTTS==2.5.0

# LangGraph for workflow orchestration
langgraph==0.0.20
langchain==0.1.4
langchain-core==0.1.16

# Audio Processing
pydub==0.25.1
mutagen==1.47.0

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.3.0

# CORS & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
